# backend/openapi.yaml
openapi: 3.0.3
info:
  title: TriLink Platform API
  description: |
    Comprehensive API documentation for the TriLink Platform - a multi-party procurement and logistics management system.
    
    ## Authentication
    Most endpoints require JWT Bearer token authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-access-token>
    ```
    
    ## Roles & Permissions
    The platform supports the following roles:
    - **Buyer**: Creates purchase requests, evaluates bids, manages contracts
    - **Supplier**: Responds to RFQs, submits bids
    - **Logistics**: Manages shipments, updates GPS tracking
    - **Clearance**: Handles customs clearance
    - **Service Provider**: Provides various services
    - **Government**: View-only access with dispute resolution capabilities
    - **Admin**: Full system access
    
    ## Company Isolation
    Most resources are isolated by company. Users can only access resources belonging to their company unless they have Admin role.
  version: 1.0.0
  contact:
    name: TriLink Platform Support
    email: support@trilink.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.trilink.com/api
    description: Production server
...
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: |
        Creates a new user account and returns access token and user data.
        Refresh token is set in httpOnly cookie (not in response body).
        Rate limited to prevent abuse.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              email: user@example.com
              password: SecurePass123!
              role: Buyer
              companyId: 507f1f77bcf86cd799439011
              firstName: John
              lastName: Doe
              phone: +971501234567
      responses:
        '201':
          description: User registered successfully
          headers:
            Set-Cookie:
              description: Refresh token set in httpOnly cookie
              schema:
                type: string
                example: refreshToken=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...; HttpOnly; Secure; SameSite=Strict; Path=/api/auth; Max-Age=604800
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      description: |
        Authenticates user and returns access token and user data.
        Refresh token is set in httpOnly cookie (not in response body).
        Rate limited to prevent brute force attacks.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: user@example.com
              password: SecurePass123!
      responses:
        '200':
          description: Login successful
          headers:
            Set-Cookie:
              description: Refresh token set in httpOnly cookie
              schema:
                type: string
                example: refreshToken=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...; HttpOnly; Secure; SameSite=Strict; Path=/api/auth; Max-Age=604800
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: |
        Generates a new access token using a valid refresh token from httpOnly cookie.
        Implements token rotation for enhanced security.
        Refresh token is automatically sent via httpOnly cookie (not in request body).
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties: {}
      responses:
        '200':
          description: Token refreshed successfully
          headers:
            Set-Cookie:
              description: New refresh token set in httpOnly cookie
              schema:
                type: string
                example: refreshToken=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...; HttpOnly; Secure; SameSite=Strict; Path=/api/auth; Max-Age=604800
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshTokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /users:
    post:
      tags:
        - Users
      summary: Create a new user
      security:
        - bearerAuth: []
      description: Creates a new user. Requires MANAGE_USERS permission.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    get:
      tags:
        - Users
      summary: Get users by company
      security:
        - bearerAuth: []
      description: Retrieves all users for a specific company. Requires VIEW_USERS permission.
      parameters:
        - $ref: '#/components/parameters/CompanyIdQuery'
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserResponse'
  /users/me:
    patch:
      tags:
        - Users
      summary: Update my profile
      security:
        - bearerAuth: []
      description: |
        Allows authenticated users to update their own profile information.
        Users can only update firstName, lastName, and phone.
        No special permissions required - users can always update their own profile.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                  minLength: 1
                  maxLength: 100
                  description: User's first name
                lastName:
                  type: string
                  minLength: 1
                  maxLength: 100
                  description: User's last name
                phone:
                  type: string
                  pattern: '^[\+]?[(]?[0-9]{1,4}[)]?[-\s\.]?[(]?[0-9]{1,4}[)]?[-\s\.]?[0-9]{1,9}[-\s\.]?[0-9]{1,9}$'
                  description: Phone number in international format
              example:
                firstName: John
                lastName: Doe
                phone: +971 50 123 4567
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseUser'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/{id}:
    get:
      tags:
        - Users
      summary: Get user by ID
      security:
        - bearerAuth: []
      description: Retrieves a specific user by ID. Requires VIEW_USERS permission.
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseUser'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - Users
      summary: Update user
      security:
        - bearerAuth: []
      description: Updates user information. Requires MANAGE_USERS permission.
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseUser'

    delete:
      tags:
        - Users
      summary: Delete user
      security:
        - bearerAuth: []
      description: Soft deletes a user. Requires MANAGE_USERS permission.
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: User deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /users/company/{companyId}:
    get:
      tags:
        - Users
      summary: Get users by company ID
      security:
        - bearerAuth: []
      description: Retrieves all users belonging to a specific company. Requires VIEW_USERS permission.
      parameters:
        - $ref: '#/components/parameters/CompanyId'
        - name: role
          in: query
          schema:
            type: string
          description: Filter by role
        - name: status
          in: query
          schema:
            type: string
          description: Filter by status
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserResponse'
...
  /payments:
    post:
      tags:
        - Payments
      summary: Create payment
      security:
        - bearerAuth: []
      description: Creates a new payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentRequest'
      responses:
        '201':
          description: Payment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePayment'

    get:
      tags:
        - Payments
      summary: Get payments
      security:
        - bearerAuth: []
      description: Retrieves payments filtered by company
      responses:
        '200':
          description: List of payments
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PaymentResponse'
  /payments/{id}:
    get:
      tags:
        - Payments
      summary: Get payment by ID
      security:
        - bearerAuth: []
      description: Retrieves a specific payment
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Payment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePayment'

    patch:
      tags:
        - Payments
      summary: Update payment
      security:
        - bearerAuth: []
      description: Updates payment information
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                paidDate:
                  type: string
                  format: date-time
                notes:
                  type: string
      responses:
        '200':
          description: Payment updated successfully

    delete:
      tags:
        - Payments
      summary: Delete payment
      security:
        - bearerAuth: []
      description: Deletes a payment
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Payment deleted successfully
...
  /payments/{id}/process:
    post:
      tags:
        - Payments
      summary: Process payment
      security:
        - bearerAuth: []
      description: Processes an approved payment
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                paymentMethod:
                  type: string
                transactionId:
                  type: string
                notes:
                  type: string
      responses:
        '200':
          description: Payment processed successfully
...
  /audit/user/{userId}:
    get:
      tags:
        - Audit
      summary: Get audit logs by user
      security:
        - bearerAuth: []
      description: Retrieves audit logs for a specific user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of audit logs
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token authentication
  parameters:
    UserId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: User ID
    CompanyId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Company ID
    CompanyIdQuery:
      name: companyId
      in: query
      schema:
        type: string
      description: Company ID filter
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Resource ID
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        error:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
        requestId:
          type: string
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            message:
              type: string
...
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - role
        - companyId
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        role:
          type: string
          enum: [Buyer, Supplier, Logistics, Clearance, Service Provider, Government, Admin]
        companyId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    # RefreshTokenRequest removed - refresh token comes from httpOnly cookie, not request body
...
    LoginResponse:
      type: object
      description: Response for login and register endpoints
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            accessToken:
              type: string
              description: JWT access token (15 minutes expiry)
          # refreshToken is NOT in response - set in httpOnly cookie
        requestId:
          type: string
    RefreshTokenResponse:
      type: object
      description: Response for refresh token endpoint
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            accessToken:
              type: string
              description: New JWT access token (15 minutes expiry)
          # refreshToken is NOT in response - new token set in httpOnly cookie
        requestId:
          type: string

    # TokenResponse removed - use RefreshTokenResponse instead

    # User schemas
    User:
      type: object
      description: User object returned in authentication responses
      properties:
        id:
          type: string
          description: User ID
        email:
          type: string
          format: email
        role:
          type: string
          enum: [Buyer, Supplier, Logistics, Clearance, Service Provider, Government, Admin]
        companyId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
    CreateUserRequest:
      type: object
      required:
        - email
        - password
        - role
        - companyId
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        role:
          type: string
        companyId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
    UpdateUserRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        status:
          type: string
          enum: [active, inactive, pending]
    UserResponse:
      type: object
      properties:
        _id:
          type: string
        email:
          type: string
        role:
          type: string
        companyId:
          type: string
        status:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        lastLogin:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ApiResponseUser:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/UserResponse'

    # Company schemas
...
    CreatePaymentRequest:
      type: object
      required:
        - contractId
        - recipientCompanyId
        - milestone
        - amount
        - dueDate
      properties:
        contractId:
          type: string
        recipientCompanyId:
          type: string
        milestone:
          type: string
        amount:
          type: number
        currency:
          type: string
          default: AED
        dueDate:
          type: string
          format: date-time
        notes:
          type: string
    PaymentResponse:
      type: object
      properties:
        _id:
          type: string
        contractId:
          type: string
        companyId:
          type: string
        recipientCompanyId:
          type: string
        buyerId:
          type: string
        milestone:
          type: string
        amount:
          type: number
        currency:
          type: string
        dueDate:
          type: string
          format: date-time
        paidDate:
          type: string
          format: date-time
        status:
          type: string
          enum: [pending_approval, approved, rejected, processing, completed, failed, cancelled, refunded]
        approvedAt:
          type: string
          format: date-time
        approvedBy:
          type: string
        rejectedAt:
          type: string
          format: date-time
        rejectedBy:
          type: string
        rejectionReason:
          type: string
        paymentMethod:
          type: string
        transactionId:
          type: string
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ApiResponsePayment:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/PaymentResponse'

    # Dispute schemas
...
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
    TooManyRequests:
      description: Too many requests - rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'